Check if the user has access to a subscription to allow create and delete board actions

+++ app/api/board/route.js
@@ -25,6 +25,13 @@ export async function POST(req) {
 
     const user = await User.findById(session.user?.id);
 
+    if (!user.hasAccess) {
+      return NextResponse.json(
+        { error: "Please subscribe first." },
+        { status: 403 }
+      );
+    }
+
     const board = await Board.create({
       userId: user._id,
       name: body.name,
@@ -57,12 +64,20 @@ export async function DELETE(req) {
       return NextResponse.json({ error: "Not authorized." }, { status: 401 });
     }
 
+    const user = await User.findById(session.user?.id);
+
+    if (!user.hasAccess) {
+      return NextResponse.json(
+        { error: "Please subscribe first." },
+        { status: 403 }
+      );
+    }
+
     await Board.deleteOne({
       _id: boardId,
       userId: session.user?.id,
     });
 
-    const user = await User.findById(session.user?.id);
     // boards is an array of boardId's
     user.boards = user.boards.filter((id) => `${id}` !== boardId);
     await user.save();
